Introduce PSR-4 Autoloader
Load Montapacking via PSR-4 autoloading
